CS-18: Fixed IDE warnings

+++ source/plugins/org.custusx.usreconstruction/cxPositionFilter.cpp
@@ -16,22 +16,22 @@ See Lisence.txt (https://github.com/SINTEFMedtek/CustusX/blob/master/License.txt
 namespace cx
 {
 
-PositionFilter::PositionFilter(int filterStrength, std::vector<class TimedPosition> &inputImagePositions) :
+PositionFilter::PositionFilter(unsigned filterStrength, std::vector<class TimedPosition> &inputImagePositions) :
     mFilterStrength(filterStrength)
 {
     mFilterLength = 1+2*filterStrength;
     mNumberInputPositions=inputImagePositions.size();
     mNumberQuaternions = mNumberInputPositions+mFilterLength;
 
-    qPosArray = Eigen::ArrayXXd::Zero(7,mNumberQuaternions);
-    qPosFiltered = Eigen::ArrayXXd::Zero(7,mNumberInputPositions);
+		qPosArray = Eigen::ArrayXXd::Zero(7,long(mNumberQuaternions));
+		qPosFiltered = Eigen::ArrayXXd::Zero(7,long(mNumberInputPositions));
 }
 
 void PositionFilter::convertToQuaternions(std::vector<class TimedPosition> &inputImagePositions)
 {
     for (unsigned int i = 0; i < mNumberQuaternions; i++) //For each pose (Tx), with edge padding
     {
-        unsigned int sourceIdx =  (i > mFilterStrength) ? (i-mFilterStrength) : 0; // Calculate index in Tx array, pad with edge elements //Skriv om
+				unsigned long sourceIdx =  (i > mFilterStrength) ? (i-mFilterStrength) : 0; // Calculate index in Tx array, pad with edge elements //Skriv om
         sourceIdx =  (sourceIdx < mNumberInputPositions) ? sourceIdx : (mNumberInputPositions-1);
         qPosArray.col(i) = matrixToQuaternion(inputImagePositions.at(sourceIdx).mPos); // Convert each Tx to quaternions
     }
@@ -41,7 +41,7 @@ void PositionFilter::filterQuaternionArray()
 {
     for (unsigned int i = 0; i < mFilterLength; i++)
     {
-        qPosFiltered = qPosFiltered + qPosArray.block(0,i,7,mNumberInputPositions);
+				qPosFiltered = qPosFiltered + qPosArray.block(0,i,7,long(mNumberInputPositions));
     }
     qPosFiltered = qPosFiltered / mFilterLength; // Scale and write back to qPosArray
 }
